inference/webui/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0

